requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
sqlalchemy==2.0.49
python-jose==3.5.0
bcrypt==5.0.0
pydantic-settings==2.9.1
slowapi==0.1.9
loguru==0.7.3
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1